+++ neo/libs/optick/optick_common.h
@@ -170,7 +170,7 @@ inline int sprintf_s(char(&buffer)[sizeOfBuffer], const char* format, ...)
 #if defined(OPTICK_GCC)
 #include <string.h>
 template<size_t sizeOfBuffer>
-inline int wcstombs_s(char(&buffer)[sizeOfBuffer], const wchar_t* src, size_t maxCount)
+inline size_t wcstombs_s(char(&buffer)[sizeOfBuffer], const wchar_t* src, size_t maxCount)
 {
 	return wcstombs(buffer, src, maxCount);
 }
@@ -178,7 +178,7 @@ inline int wcstombs_s(char(&buffer)[sizeOfBuffer], const wchar_t* src, size_t ma
 
 #if defined(OPTICK_MSVC)
 template<size_t sizeOfBuffer>
-inline int wcstombs_s(char(&buffer)[sizeOfBuffer], const wchar_t* src, size_t maxCount)
+inline size_t wcstombs_s(char(&buffer)[sizeOfBuffer], const wchar_t* src, size_t maxCount)
 {
 	size_t converted = 0;
 	return ::wcstombs_s(&converted, buffer, src, maxCount);
